refs piwik/piwik#5864 do not commit changes if generation fails but output error to STDERR

+++ generateAndPush.sh
@@ -9,6 +9,11 @@ cd ..
 git checkout docs/generated
 git pull
 ./generate.sh
+GENERATION_SUCCESS=$?
+if [ $GENERATION_SUCCESS -ne 0 ]; then
+  echo -e "Failed to generate documentation" 1>&2;
+  exit 1;
+fi
 git add docs/generated
 git rm $(git ls-files --deleted docs/generated)
 git commit -m 'updated plugins API documentation'
